Change github post methods to make changes backwards compatible.
This will mean that these methods work whether the pipeline calling them
use the commit hash or not. This allows us to change the method whilst
ensuring it does not break many Jenkins pipelines.

Once this change is merged we can then change the individual Jenkins
pipelines then come back and remove the null functionality from the
affected methods.