Copy library straight from target directory to S3
Remove intermediate step where the library is copied from the Scala
target directory to the root directory. Copying to the root directory
doesn't work any more because the Jenkins nodes now run as the jenkins
user rather than the root user, and the jenkins user doesn't have
permission to copy files to the root directory.