Add beforeAgent parameter.
Jenkins is trying to spin up the container which deploys the lambda
before evaluating whether or not the branch is main/master.

But the original Jenkins can no longer assume the role for things that
run in the management account, so things like the notifications multi
branch pipeline fail.

The beforeAgent https://www.jenkins.io/doc/book/pipeline/syntax/#evaluating-when-before-entering-agent-in-a-stage
parameter means that the expression is checked before the agent started.
This will solve this problem and will speed up other builds as we're not
waiting for an ECS task to start just to be skipped because the code
isn't on the master/main branch.