Add release to github step.
This is part of the work to have a backend check performance test on
the sandbox environment. It will be easier to get the code for the
lambdas if it is deployed publicly rather than trying to allow access to
the management bucket for a sandbox account.
There's nothing sensitive stored in these jars so they're ok to be
released publicly.

This could be added in as part of the sbt build process like we do with
the export but we have python lambdas as well and I wanted the process
to work for everything.